feat: add airflow Docker deployment files
This commit adds the necessary Docker deployment files for the
Airflow project. This includes a docker-compose file for the Airflow
service, a service account config file for authentication, and updates
to the .gitignore file to ignore IDE specific files. This setup
allows for a seamless deployment of the Airflow service with all its
dependencies.